Edit pass for Connectors Gallery DO NOT MERGE
Light copyediting of the Connectors Gallery landing page. The spelling errors called out in the work item were not present, so this article might have already recieved an edit.

Edits were mostly punctuation, removing double-spaces, and a few minor edits for clarity.

One question for the author: in the LDAP by BA Insight entry, there was a sentence that read "....making it simple to search for users, contract, or groups stored anywhere in your directory."

I assumed that "contract" was meant to be "contacts", and edited accordingly. Please verify that this is correct.