Correct links to "Customize the search results page"
It seems like the links to "Customize the search results page" were probably bad for a while, since they refer to a section in configure-connector.md, and I didn't see that heading in any recent version. This link was revealed as bad thanks to the script that changed the link from absolute to relative (PR https://github.com/MicrosoftDocs/OfficeDocs-MicrosoftSearch-pr/pull/540).