Fixed "[Suggestion-code-block-indented]"
This single-line example was presented as a code block due to its indentation in Markdown. However, since it has double asterisks around it, I assume this was meant to be in bold. Since the asterisks aren't part of the code, I've removed the indentation so that it is displayed as bold text with the asterisks being visible. Another solution would be to remove the asterisks and put this line inside of two other lines, each of which has triple backtick characters.