Targeted Release for Preview Connectors
We need to document that customer must be on Targeted Release to see the Preview Connectors. I don't know where this should go but it should go on this set up document as another note, as well as for all connectors that are in Preview. Please see Vivek Gupta for more information.