Update manage-acronyms.md
Hi,

I wanted to bring note to this exception when there is a space in the acronym string. The form does not return a validation error and the acronym is listed under the section however searches for the string that includes a space will not return the desired acronym answer. 

Thanks.