Update confluence-onpremises-connector.md
PR updated based on comments.
Line 11: Replace localization_priority with ms.localizationpriority to correct the metadata field.
Line 69, 73, 74: Start the alt text with graphic type such as “Screenshot that shows…” or “Diagram of…” and revise the alt text to adequately describe the image content. Also, add lightbox to improve the readability of images.
remove this line no 72 [ Make a call to the URL: http://{BASE_URL}/rest/api/msplugin/1.0/healthcheck/version and check if it gives you the version. If yes, you are all set for crawling]
Add a Note! -
Plugin is supported for confluence version above 8.0