Update troubleshoot-confluence-onpremises-connector.md
Line 11: Replace localization_priority with ms.localizationpriority to correctly represent the metadata field.
Line 32, 33: Both the headers look weird on staging the content. Put them in the form of paragraph, if possible, so that the content looks clean on staging.
Line 36: Replace Note with [!NOTE] and put the content as a Markdown alert or use the phrase "Note that" to start the sentence.